+++ posthog-react-native/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Removed
 
-1. remove `captureMode` in favor or `json` capture mode only
+1. Remove `captureMode` in favor or `json` capture mode only
+2. Remove deprecated `personProperties` and `groupProperties` in favor of `setPersonPropertiesForFlags` and `setGroupPropertiesForFlags`
+3. Rename `captureNativeAppLifecycleEvents` option to `captureAppLifecycleEvents`
+    1. `captureAppLifecycleEvents` from `autocapture` is removed and replaced by `captureAppLifecycleEvents` from options
+4. Removed `version` and `build` from Lifecycle events in favor of `$app_version` and `$app_build`
 
 # 3.15.1 – 2025-04-28